requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
aiosqlite==0.18.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
discord==2.1.0
discord.py==2.1.0
frozenlist==1.3.3
idna==3.4
iota-client-python # @ file:///iota_client_python-0.2.0a3-cp36-abi3-linux_x86_64.whl https://nightly.link/iotaledger/iota.rs/workflows/python-bindings-publish/production
multidict==6.0.4
requests==2.28.2
urllib3==1.26.14
yarl==1.8.2
